pyproject.toml
[tool.poetry]
name = "Mini-Game_Discord_Bot"
version = "0.1.0"
description = ""
authors = ["RaptorIndian <raptorindian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^1.7.3"
"discord.py" = "^1.7.3"
poetry-dotenv-plugin = "^0.1.0-alpha.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

packages = [
    { include = "game.py", from = "src" },
]